requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3
phonenumbers==8.13.22
pycountry==23.12.11
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2
httpx==0.25.2
